+++ src/main/java/personal/social/repository/CommentRepository.java
@@ -1,9 +1,17 @@
 package personal.social.repository;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import personal.social.dto.CommentDTO;
 import personal.social.model.Comment;
 
+import java.util.List;
+
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
 }
