+++ src/main/java/personal/social/repository/PostContentRepository.java
@@ -7,5 +7,5 @@ import personal.social.model.Post;
 import personal.social.model.PostContent;
 
 @Repository
-public interface PostContentRepository extends JpaRepository<PostContent, Integer> {
+public interface PostContentRepository extends JpaRepository<PostContent, Long> {
 }
