+++ src/main/java/personal/social/repository/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Repository;
 import personal.social.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByEmail(String email);
     User findByPhone(String phone);
-    User findByUserId (Integer userId);
+    User findByUserId (Long userId);
 }
